Update app.js
We define the getUserChoice() function, which loops through the answer choice radio buttons to identify the choice selected by the user. This function updates the userChoice variable with the selected answer choice and returns true if there was a selected answer, or displays an alert to the user and returns false otherwise.
We define the gradeQuestion() function, which compares the user choice (if selected) to the correct answer for the current question. This function updates the score variable if the answer was correct, and stores the appropriate value in the scores array. It then updates the current variable to point to the next question and calls the renderQuestion() function to display the next question.